some renaming and moving
