+++ src/spatialregistry/RTree.java
@@ -1,4 +1,4 @@
-package spatialRegistries;
+package spatialregistry;
 
 import java.util.Arrays;
 import java.util.HashSet;
